Add a vimtutor script
It would be interesting for w64devkit to be part of some sort of intro
to programming that targets Windows users who haven't yet graduated to a
more appropriate operating system. These people would need vimtutor in
order to make use of Vim.

Unfortunately the Vim developers decided that vimtutor should be a
complicated, fragile script rather than a feature built into Vim. After
a couple hours of trying, I still can't get the original vimtutor.bat
working reliably. So instead I included a one liner script that does 95%
of the job reliably. Unfortunately it's not an exact match with the
instructions, but this is the best I've found so far.