WIP Add -k (terminate) option to debugbreak
Unfortunately this does not work as well as I had hoped. If the debugee
is currently paused in a debug event, it will not be terminated until it
continues. The whole point of this tool is to be quick, convenient, and
thoughtless, but mentally tracking the debugee state is extra cognitive
load. Might as well just get in the habit of killing the process from
the debugger, even if it requires running debugbreak first.