Add "title" configuration to w64devkit.ini
While already possible using an ANSI escape sequence, the INI option
affords additional flexibility. It also supports wide titles, which
cannot be done with the busybox-w32 included in w64devkit. See #115.

To support this, I broke up homeconfig() into several functions to be
re-used and composed when handling the new option, such as expanding
environment variables. While doing so, I took the opportunity to remove
several maximum length limitations, like MAX_PATH, and so long paths are
now supported, at least as far as the launcher. I also simplified the
arena to a pair of pointers.

The program now sets $W64DEVKIT before loading the INI so that it is
available during variable expansion of titles.