Enable libbacktrace in libstdc++
This was supposed to be automatic, but the libstdc++ configure script
incorrectly applies libbacktrace's filetype.awk against an image (.exe)
instead of an object file (.o). This works out by accident on ELF hosts
because it's the same format. It overrides --enable-stdcxx-backtrace,
and so even force-enabling it has no effect. To mitigate the defect,
clobber libbacktrace just before cross-compiling GCC.

As of GCC 14.2, <stacktrace> requires linking -lstdc++exp.

Fixes #200.